requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-filter==21.1
drf-nested-routers
django-rest-framework-nested